aws/ec2-app/requirements.txt
requests==2.31.0
boto3==1.28.57
pyjwt[crypto]==2.8.0
cryptography==41.0.7
python-dotenv==0.21.1
urllib3<2.0.0
